debugging

+++ src/main/java/miniplc0java/tokenizer/Tokenizer.java
@@ -190,7 +190,7 @@ public class Tokenizer {
         else if(peek!='\'' && peek!='\n' && peek!='\t' && peek!='\r'){
             char c=it.nextChar();
             if(it.nextChar()=='\'')
-                return new Token(TokenType.CHAR_LITERAL, (long)c, start, it.currentPos());
+                return new Token(TokenType.CHAR_LITERAL, c, start, it.currentPos());
             throw new TokenizeError(ErrorCode.InvalidInput, start);
         }
         throw new TokenizeError(ErrorCode.InvalidInput, start);
